requirements.txt
streamlit>=1.36.0
openai>=1.40.0
pyyaml>=6.0.1
pandas>=2.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
